% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apiData.R
\name{apiData}
\alias{apiData}
\title{Query data from an API}
\usage{
apiData(x, shortnames=FALSE)
}
\arguments{
\item{x}{an apirequest object}

\item{shortnames}{logical, indicating whether the resulting tables (data frames) should have short variable names (default is FALSE, variable names contain nesting hierarchy)}

\item{method}{character, either "RwebAPI", "jsonlite", or "XML2R"}
}
\value{
an apidata-object containing the returned data in a flat representation
}
\description{
A high-level function that automates the querying and extraction of data from a web API.
}
\examples{
\dontrun{apidata <- apiData(x) }
}

