% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apiDatalight.R
\name{apiDatalight}
\alias{apiDatalight}
\title{Query data from an API}
\usage{
apiDatalight(x, ...)
}
\arguments{
\item{x}{an apirequest object}

\item{...}{currently only one parameter (simplify) passed down to the mapping algorithm if simplify is TRUE, the document tree is made simpler if possible (by removing unnecessary nodes)}
}
\value{
a list containing the returned data in a flat representation
}
\description{
A high-level function that automates the querying and extraction of data from a web API. Unlike apiData(), only the pure data is returned.
}
\examples{
\dontrun{apidata <- apiDatalight(x)}
}

