% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apiEasyData.R
\name{apiEasyData}
\alias{apiEasyData}
\title{Directly interact with any API}
\usage{
apiEasyData(x, base.url, shortnames=TRUE, ...)
}
\arguments{
\item{x}{either a charachter string containing the whole url for the request or a named list containing the parameter names and values (see details)}

\item{base.url}{a character string containing the basic url for the api}

\item{shortnames}{logical, indicating whether the resulting tables (data frames) should have short variable names (default is FALSE, variable names contain nesting hierarchy)}

\item{...}{currently only one parameter (simplify) passed down to the mapping algorithm if simplify is TRUE, the document tree is made simpler if possible (by removing unnecessary nodes)}
}
\value{
a list of data-frames, containing the returned data in a flat representation
}
\description{
A high-level function that directly queries and transforms data from any web API.
}
\examples{
\dontrun{apidata <- apiEasyData(x) }
}

