% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Extract API Data}
\usage{
getdata(x)
}
\arguments{
\item{x}{an object of class apidata}
}
\value{
either a data frame or a list containing data frames.
}
\description{
Extracts the converted data from an apidata object.
}
\examples{
\dontrun{apidata <- apiData(x) # only works with a proper PVS API key}
\dontrun{getdata(apidata)}
}
