% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTabularData.R
\name{getTabularData}
\alias{getTabularData}
\title{Get web data in tabular form (data frames)}
\usage{
getTabularData(x, base.url, shortnames=TRUE, alignVariables=FALSE)
}
\arguments{
\item{x}{either a charachter string containing the whole url for the request or a named list 
containing the parameter names and values (see details)}

\item{base.url}{a character string containing the basic url for the api}

\item{shortnames}{logical, indicating whether the resulting tables (data frames) should have 
short variable names (default is FALSE, variable names contain nesting hierarchy)}

\item{alignVariables}{logical, indicating whether variables/values should be rearranged in case the raw data was malformed (missing variable names)}
}
\value{
a list of data-frames, containing the returned data in a flat representation
}
\description{
A high-level function that directly queries and transforms data from any web API to a data frame or list of data frames.
}
\examples{
\dontrun{apidata <- getTabularData(x) }
}

