% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.apidata.R
\name{is.apidata}
\alias{is.apidata}
\title{Check if object is an apidata}
\usage{
is.apidata(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical, TRUE if x is of class apidata.
}
\description{
A function to check if an object is of class "apidata".
}

