% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.apirequest.R
\name{is.apirequest}
\alias{is.apirequest}
\title{Check if object is an apirequest}
\usage{
is.apirequest(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical, TRUE if x is of class apirequest.
}
\description{
A function to check if an object is of class "apisreqest".
}
