% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.apiresp.R
\name{is.apiresp}
\alias{is.apiresp}
\title{Check if object is an apiresp}
\usage{
is.apiresp(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical, TRUE if x is of class apiresp.
}
\description{
A function to check if an object is of class "apiresp".
}

