% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.apiresponse.R
\name{is.apiresponse}
\alias{is.apiresponse}
\title{Check if object is an apiresponse}
\usage{
is.apiresponse(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical, TRUE if x is of class apiresponse.
}
\description{
A function to check if an object is of class "apiresponse".
}
