% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapiDatalight.R
\name{mapiDatalight}
\alias{mapiDatalight}
\title{Query data from an API}
\usage{
mapiDatalight(x, ...)
}
\arguments{
\item{x}{list of apirequest objects or urls}

\item{...}{passed on to apiDatalight()}
}
\value{
a (nested) list containing the returned data in a flat representation
}
\description{
A vectorized version of apiDatalight(), wrapped around apiDatalight()
}
\examples{
\dontrun{apidata <- mapiDatalight(list(mr,mr2)) # only works with a proper PVS API key}
}

