% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redlist.R
\name{redlist}
\alias{redlist}
\title{Reduce a nested list}
\usage{
redlist(x)
}
\arguments{
\item{x}{a nested list containing some data frames}
}
\value{
a list containing data frames
}
\description{
A function that reduces a nested list containing some data frames in some of its elements (a list containing lists containing lists containing data frames...) by unlisting its elements (lists) until it only consits of a normal list containing the dataframes as elements.
}
\examples{
a <- data.frame(a=c(1,2,4),b=c(5,5,6))
b <- data.frame(a=c(3,4,5,6), b=c(8,8,5,4), c=c(9,4,2,1))
l1 <- list(list(list(a,b)))
redlist(x=l1)
}

