% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAPIkey.R
\name{saveAPIkey}
\alias{saveAPIkey}
\title{Save API keys}
\usage{
saveAPIkey(key.var, key)
}
\arguments{
\item{key.var}{a character string with the variable name the key should be saved in.}

\item{key}{the api-key as character string}
}
\value{
either one  data frame or a list containing several data.frames into which the tree structured web-data has been transformed.
}
\description{
A function that saves the api key as character in a specific variable in the API keys environment.
}
\details{
The name of the environment the key is assigned to defaults to "apikeys".
}
\examples{
saveAPIkey(key.var="disq", key="1234" )
get("disq", pos="apikeys")
}

