% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/XMLtoDataFrame.R
\name{XMLtoDataFrame}
\alias{XMLtoDataFrame}
\title{Extract the data from a XML document as one (or several) data frame(s)}
\usage{
XMLtoDataFrame(x, alignVariables=FALSE)
}
\arguments{
\item{x}{a string with the path of a XML document}

\item{alignVariables}{logical, indicating whether variables/values should be rearranged in case the raw data was malformed (missing variable names)}
}
\value{
one data frame or a list of several data frames
}
\description{
Given a XML document, this function maps the potentially nested XML data to one or several data frames.
The data extraction algorithm applied in this function is based on the assumption that the XML document describes either one or several entity types containing a set of observations described by different variables.
}
\details{
The data extraction algorithm applied in this function partly relies on a nested (tree-structured) data representation \cr
         it is favorable for element based XML. XML documents that largely build on attributes can also be processed. \cr
         Several attributes of the same tag will, however, be collected in the same data-frame column.
}
\examples{
XML.ex <- system.file("exdata", "XML_fiction.xml", package = "RWebData")
XMLtoDataFrame(XML.ex, alignVariables=FALSE)
}

