% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YAMLtoDataFrame.R
\name{YAMLtoDataFrame}
\alias{YAMLtoDataFrame}
\title{Extract the data from a YAML document as one (or several) data frame(s)}
\usage{
YAMLtoDataFrame(x, alignVariables=FALSE)
}
\arguments{
\item{x}{either a string containing YAML or the name of a file containing the YAML}

\item{alignVariables}{logical, indicating whether variables/values should be rearranged in case the raw data was malformed (missing variable names)}
}
\value{
one data frame or a list of several data frames
}
\description{
Given YAML document, this function maps the potentially nested YAML data to one or several data frames.
The data extraction algorithm applied in this function is based on the assumption that the YAML document describes either one or several entity types containing a set of observations described by different variables.
}
\examples{
yaml.ex <- system.file("exdata", "microcapital.yml", package = "RWebData")
YAMLtoDataFrame(yaml.ex, alignVariables=FALSE)
}
